+++ core/rend/vulkan/oit/oit_drawer.h
@@ -87,7 +87,7 @@ protected:
 	{
 		if (mainBuffers.empty())
 		{
-			for (int i = 0; i < GetContext()->GetSwapChainSize(); i++)
+			for (size_t i = 0; i < GetContext()->GetSwapChainSize(); i++)
 				mainBuffers.push_back(std::unique_ptr<BufferData>(new BufferData(std::max(512 * 1024u, size),
 						vk::BufferUsageFlagBits::eVertexBuffer | vk::BufferUsageFlagBits::eIndexBuffer | vk::BufferUsageFlagBits::eUniformBuffer
 						| vk::BufferUsageFlagBits::eStorageBuffer)));
